requirements.txt
Django==1.11.13
psycopg2-binary==2.8.4
cx_Oracle==7.3.0
django-dotenv==1.4.2
dj-database-url==0.5.0
